app/requirements.txt
streamlit>=1.36
openai>=1.40
pydantic>=2.6
python-dotenv>=1.0
